refactor(portal): extract messaging JS into separate module files
Move the inline JavaScript out of portal/messaging/messages.php and
secure_chat.php into portal/messaging/js/messages.js and secure_chat.js.
The PHP files now emit a small window.OE_MESSAGES_CONFIG /
OE_SECURE_CHAT_CONFIG block (translations, session-derived user
identity, mode flags, initial inbox payload, CSRF token) and pull the
static JS via a versioned <script src>.

Benefits:
- ESLint and code-review tools can now see the JS clearly. Prior to
  this the messaging UI was ~1500 lines of inline JS spread between two
  PHP heredocs with 22 PHP injection sites that escape-context tools
  couldn't reason about.
- The SMS event-dispatch hook
  (SendSmsEvent::JAVASCRIPT_READY_SMS_POST) is now emitted at the
  module's top level instead of inside the IIFE; this also fixes a
  latent issue where its `function sendSMS(pid, phone)` definition was
  scoped to the IIFE and unreachable from button onclicks.
- Adds portal/messaging/js/ to codecov ignore so the UI code, which is
  exercised by manual QA rather than unit tests, doesn't drag
  patch-coverage on every PR touching it.

No behavior change to the UI logic itself; this is a move + adapt:
every PHP-injected translation/value is now read from the
OE_*_CONFIG.strings object that PHP populates before the static JS
loads. Verified PHPCS/PHPStan/ESLint clean.
