+++ codecov.yml
@@ -3,10 +3,11 @@ ignore:
 # into this path by the oe-module-installer-plugin during `composer install`.
 # Third-party code, not maintained in this repo — exclude from coverage.
 - interface/modules/custom_modules/oe-module-claimrev-connect/
-- config/                # Container/service configuration — not directly testable
-- tests/PHPStan/         # Static analysis rules — not exercised by PHPUnit
-- tests/Rector/          # Rector rules — run outside PHPUnit coverage
-- tests/eventdispatcher/ # Example modules for documentation
+- config/                  # Container/service configuration — not directly testable
+- tests/PHPStan/           # Static analysis rules — not exercised by PHPUnit
+- tests/Rector/            # Rector rules — run outside PHPUnit coverage
+- tests/eventdispatcher/   # Example modules for documentation
+- portal/messaging/js/     # Patient portal messaging UI — covered by manual QA, not unit tests
 
 # Strip environment-specific path prefixes from coverage reports so Codecov
 # can match them to repository files. See https://docs.codecov.com/docs/fixing-paths
