+++ portal/messaging/messages.php
@@ -128,746 +128,38 @@ function getAuthPortalUsers()
     <meta name="description" content="Mail Application" />
 </head>
 <body class="body_top">
+<?php
+$webRoot = $globalsBag->getString('web_root');
+$jsVersion = $globalsBag->getString('v_js_includes', '');
+?>
     <script>
-        (function () {
-            'use strict';
-
-            // --- State ---
-            const state = {
-                sortingOrder: 'id',
-                reverse: false,
-                filteredItems: [],
-                itemsPerPage: 20,
-                pagedItems: [],
-                compose: { task: 'add' },
-                selrecip: [],
-                currentPage: 0,
-                sentItems: [],
-                allItems: [],
-                deletedItems: [],
-                inboxItems: <?php echo json_encode($theresult);?>,
-                items: [],
-                selected: null,
-                userproper: <?php echo !empty($session->get('ptName', null)) ? js_escape($session->get('ptName')) : js_escape($dashuser['fname'] . ' ' . $dashuser['lname']);?>,
-                isPortal: <?php echo json_encode(IS_PORTAL ?: ''); ?>,
-                isDashboard: <?php echo json_encode(IS_DASHBOARD ?: ''); ?>,
-                authrecips: <?php echo json_encode(getAuthPortalUsers());?>,
-                xLate: {
-                    fwd: <?php echo xlj('Forwarded Portal Message Re: '); ?>,
-                    confirm: {
-                        one: <?php echo xlj('Confirm to Archive Current Thread?'); ?>,
-                        all: <?php echo xlj('Confirm to Archive Selected Messages?'); ?>,
-                        err: <?php echo xlj('You are sending to yourself!'); ?>
-                    }
-                },
-                csrf: <?php echo js_escape(CsrfUtils::collectCsrfToken($session, 'messages-portal')); ?>,
-                isInit: false,
-                isInbox: true,
-                isSent: false,
-                isAll: false,
-                isTrash: false,
-                errorLoadingMessages: false
-            };
-
-            state.cUserId = state.isPortal || state.isDashboard;
-            state.items = state.inboxItems;
-
-            // --- Helpers ---
-
-            function escapeHtml(text) {
-                if (text === null || text === undefined) return '';
-                const div = document.createElement('div');
-                div.textContent = String(text);
-                return div.innerHTML;
+        window.OE_MESSAGES_CONFIG = {
+            inboxItems: <?php echo json_encode($theresult); ?>,
+            userproper: <?php echo !empty($session->get('ptName', null)) ? js_escape($session->get('ptName')) : js_escape($dashuser['fname'] . ' ' . $dashuser['lname']); ?>,
+            isPortal: <?php echo json_encode(IS_PORTAL ?: ''); ?>,
+            isDashboard: <?php echo json_encode(IS_DASHBOARD ?: ''); ?>,
+            authrecips: <?php echo json_encode(getAuthPortalUsers()); ?>,
+            csrf: <?php echo js_escape(CsrfUtils::collectCsrfToken($session, 'messages-portal')); ?>,
+            strings: {
+                forwardedRe: <?php echo xlj('Forwarded Portal Message Re: '); ?>,
+                confirmOne: <?php echo xlj('Confirm to Archive Current Thread?'); ?>,
+                confirmAll: <?php echo xlj('Confirm to Archive Selected Messages?'); ?>,
+                sendingToSelf: <?php echo xlj('You are sending to yourself!'); ?>,
+                conversationFrom: <?php echo xlj('Conversation from'); ?>,
+                regarding: <?php echo xlj('regarding'); ?>,
+                onPrep: <?php echo xlj('on'); ?>,
+                archiveTitle: <?php echo xlj('Archive this message'); ?>,
+                composeReplyTitle: <?php echo xlj('Compose Reply Message'); ?>,
+                composeNewTitle: <?php echo xlj('Compose New Message'); ?>
             }
-
-            function searchMatch(haystack, needle) {
-                if (!needle) return true;
-                return String(haystack).toLowerCase().indexOf(needle.toLowerCase()) !== -1;
-            }
-
-            function renderMessageBody(html) {
-                return DOMPurify.sanitize(html, {
-                    USE_PROFILES: { html: true },
-                    FORBID_TAGS: ['a', 'img']
-                });
-            }
-
-            function htmlToText(html) {
-                const hold = document.createElement('DIV');
-                hold.innerHTML = DOMPurify.sanitize(html, {
-                    USE_PROFILES: { html: true },
-                    FORBID_TAGS: ['a', 'img']
-                });
-                return hold.textContent || hold.innerText || '';
-            }
-
-            function limitTo(str, limit) {
-                if (!str) return '';
-                return str.length > limit ? str.substring(0, limit) : str;
-            }
-
-            async function postForm(url, params) {
-                const response = await fetch(url, {
-                    method: 'POST',
-                    credentials: 'same-origin',
-                    headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-                    body: params
-                });
-                return response.json();
-            }
-
-            // --- Filtering & Paging ---
-
-            function search(query) {
-                if (!query) {
-                    state.filteredItems = state.items.slice();
-                } else {
-                    state.filteredItems = state.items.filter(item => {
-                        for (const attr in item) {
-                            if (searchMatch(item[attr], query)) return true;
-                        }
-                        return false;
-                    });
-                }
-                state.currentPage = 0;
-                groupToPages();
-            }
-
-            function groupToPages() {
-                state.selected = null;
-                state.pagedItems = [];
-                for (let i = 0; i < state.filteredItems.length; i++) {
-                    const pageIdx = Math.floor(i / state.itemsPerPage);
-                    if (i % state.itemsPerPage === 0) {
-                        state.pagedItems[pageIdx] = [state.filteredItems[i]];
-                    } else {
-                        state.pagedItems[pageIdx].push(state.filteredItems[i]);
-                    }
-                }
-                renderAll();
-            }
-
-            function prevPage() {
-                if (state.currentPage > 0) {
-                    state.currentPage--;
-                    renderAll();
-                }
-            }
-
-            function nextPage() {
-                if (state.currentPage < state.pagedItems.length - 1) {
-                    state.currentPage++;
-                    renderAll();
-                }
-            }
-
-            // --- Folder selection ---
-
-            function selectInbox() {
-                state.isInbox = true;
-                state.isTrash = state.isAll = state.isSent = false;
-                state.items = state.inboxItems;
-                search();
-            }
-
-            function selectSent() {
-                state.isSent = true;
-                state.isTrash = state.isAll = state.isInbox = false;
-                state.items = state.sentItems;
-                search();
-            }
-
-            function selectAll() {
-                state.isAll = true;
-                state.isTrash = state.isSent = state.isInbox = false;
-                state.items = state.allItems;
-                search();
-            }
-
-            function selectTrash() {
-                state.isTrash = true;
-                state.isSent = state.isAll = state.isInbox = false;
-                state.items = state.deletedItems;
-                search();
-            }
-
-            // --- API calls ---
-
-            async function getAllMessages() {
-                try {
-                    const data = await postForm('handle_note.php', $.param({ task: 'getall', csrf_token_form: state.csrf }));
-                    if (data) {
-                        state.allItems = Array.isArray(data) ? data.slice() : [];
-                    }
-                } catch (e) {
-                    console.error('Error getting all messages:', e);
-                }
-            }
-
-            async function getDeletedMessages() {
-                try {
-                    const data = await postForm('handle_note.php', $.param({ task: 'getdeleted', csrf_token_form: state.csrf }));
-                    if (data) {
-                        state.deletedItems = Array.isArray(data) ? data.slice() : [];
-                    }
-                } catch (e) {
-                    console.error('Error getting deleted messages:', e);
-                }
-            }
-
-            async function getSentMessages() {
-                try {
-                    const data = await postForm('handle_note.php', $.param({ task: 'getsent', csrf_token_form: state.csrf }));
-                    state.sentItems = Array.isArray(data) ? data.slice() : [];
-                } catch (e) {
-                    console.error('Error getting sent messages:', e);
-                }
-            }
-
-            async function deleteMessage(id) {
-                try {
-                    await postForm('handle_note.php', $.param({ task: 'delete', noteid: id, csrf_token_form: state.csrf }));
-                } catch (e) {
-                    alert(e.message);
-                    throw e;
-                }
-            }
-
-            function deleteItem(item) {
-                if (!item) return;
-                if (!confirm(state.xLate.confirm.one)) return;
-                const removeFromList = (list) => {
-                    if (!Array.isArray(list)) return;
-                    for (let i = list.length - 1; i >= 0; i--) {
-                        if (list[i] && list[i].mail_chain === item.mail_chain) {
-                            list.splice(i, 1);
-                        }
-                    }
-                };
-                deleteMessage(item.mail_chain).then(() => {
-                    removeFromList(state.items);
-                    removeFromList(state.inboxItems);
-                    removeFromList(state.sentItems);
-                    removeFromList(state.allItems);
-                    if (state.selected && state.selected.mail_chain === item.mail_chain) {
-                        state.selected = null;
-                    }
-                    search();
-                    getDeletedMessages();
-                });
-            }
-
-            function batchDelete() {
-                if (!confirm(state.xLate.confirm.all)) return;
-                const itemToDelete = [];
-                state.items.forEach((item, idx) => {
-                    if (item._deleted) {
-                        itemToDelete.push(item.id);
-                    }
-                });
-                postForm('handle_note.php', $.param({
-                    task: 'massdelete',
-                    notejson: JSON.stringify(itemToDelete),
-                    csrf_token_form: state.csrf
-                })).then(() => {
-                    window.location.reload();
-                }).catch(e => {
-                    alert(e.message);
-                });
-            }
-
-            function readMessage(item) {
-                if (!item) return;
-                if (item.message_status === 'New') {
-                    postForm('handle_note.php', $.param({
-                        task: 'setread',
-                        noteid: item.id,
-                        csrf_token_form: state.csrf
-                    })).then(() => {
-                        const markReadById = (list, id) => {
-                            if (!Array.isArray(list)) return;
-                            list.forEach(entry => {
-                                if (entry && entry.id === id) {
-                                    entry.message_status = 'Read';
-                                }
-                            });
-                        };
-                        markReadById(state.items, item.id);
-                        markReadById(state.inboxItems, item.id);
-                        markReadById(state.sentItems, item.id);
-                        markReadById(state.allItems, item.id);
-                        renderAll();
-                    });
-                }
-                state.selected = item;
-                renderAll();
-            }
-
-            function closeMessage() {
-                state.selected = null;
-                groupToPages();
-            }
-
-            function readAll() {
-                state.items.forEach(item => { item.message_status = 'Read'; });
-                renderAll();
-            }
-
-            // --- Rendering ---
-
-            function renderAll() {
-                renderNav();
-                renderToolbar();
-                renderMessageList();
-                renderMessageDetail();
-                renderPaging();
-                renderLoadingState();
-            }
-
-            function renderLoadingState() {
-                const loadingRow = document.getElementById('loadingRow');
-                const contentRow = document.getElementById('contentRow');
-                if (loadingRow && contentRow) {
-                    if (state.isInit) {
-                        loadingRow.classList.add('d-none');
-                        contentRow.classList.remove('d-none');
-                    } else {
-                        loadingRow.classList.remove('d-none');
-                        contentRow.classList.add('d-none');
-                    }
-                }
-            }
-
-            function renderNav() {
-                document.querySelectorAll('#navInbox, #navSent, #navAll, #navTrash').forEach(el => {
-                    el.classList.remove('active');
-                });
-                if (state.isInbox) document.getElementById('navInbox').classList.add('active');
-                if (state.isSent) document.getElementById('navSent').classList.add('active');
-                if (state.isAll) document.getElementById('navAll').classList.add('active');
-                if (state.isTrash) document.getElementById('navTrash').classList.add('active');
-
-                const inboxBadge = document.getElementById('badgeInbox');
-                const sentBadge = document.getElementById('badgeSent');
-                const allBadge = document.getElementById('badgeAll');
-                const trashBadge = document.getElementById('badgeTrash');
-                if (inboxBadge) inboxBadge.textContent = state.inboxItems.length;
-                if (sentBadge) sentBadge.textContent = state.sentItems.length;
-                if (allBadge) allBadge.textContent = state.allItems.length;
-                if (trashBadge) trashBadge.textContent = state.deletedItems.length;
-
-                // Hide archive option when viewing the archive folder
-                const archiveMenuItem = document.getElementById('archiveMenuItem');
-                if (archiveMenuItem) {
-                    archiveMenuItem.style.display = state.isTrash ? 'none' : '';
-                }
-            }
-
-            function renderToolbar() {
-                const toolbar = document.getElementById('inboxToolbar');
-                if (toolbar) {
-                    toolbar.style.display = state.selected ? 'none' : '';
-                }
-            }
-
-            function renderMessageList() {
-                const container = document.getElementById('messageListTable');
-                if (!container) return;
-                container.style.display = state.selected ? 'none' : '';
-
-                const tbody = container.querySelector('tbody');
-                if (!tbody) return;
-
-                const currentPageItems = state.pagedItems[state.currentPage] || [];
-                let html = '';
-                currentPageItems.forEach((item, idx) => {
-                    const strong = item.message_status === 'New' ? 'font-weight: bold;' : '';
-                    const bodyPreview = limitTo(htmlToText(item.body), 35);
-                    const dateStr = item.date || '';
-                    html += `<tr role="button">
-                        <td class="message-row">
-                            <span class="col-sm-1" style="max-width: 5px;"><input type="checkbox" data-item-idx="${idx}" class="item-checkbox"></span>
-                            <span class="col-sm-1 px-1 msg-click" data-idx="${idx}"><span style="${strong}">${escapeHtml(item.message_status)}</span></span>
-                            <span class="col-sm-2 px-1 msg-click" data-idx="${idx}"><span style="${strong}">${escapeHtml(dateStr)}</span></span>
-                            <span class="col-sm-3 px-1 msg-click" data-idx="${idx}">
-                                <a class="btn-link"><span style="${strong}">${escapeHtml(item.sender_name)} to ${escapeHtml(item.recipient_name)}</span></a>
-                            </span>
-                            <span class="col-sm-1 msg-click" data-idx="${idx}">
-                                <a class="btn-link"><span style="${strong}">${escapeHtml(item.title)}</span></a>
-                            </span>
-                            <span class="col-sm-4 px-1 msg-click" data-idx="${idx}"><span style="${strong}">${escapeHtml(bodyPreview)}</span></span>
-                        </td>
-                    </tr>`;
-                });
-                tbody.innerHTML = html;
-
-                // Bind click handlers
-                tbody.querySelectorAll('.msg-click').forEach(el => {
-                    el.addEventListener('click', () => {
-                        const idx = parseInt(el.dataset.idx, 10);
-                        const item = (state.pagedItems[state.currentPage] || [])[idx];
-                        readMessage(item);
-                    });
-                });
-                tbody.querySelectorAll('.item-checkbox').forEach(cb => {
-                    cb.addEventListener('change', function () {
-                        const idx = parseInt(this.dataset.itemIdx);
-                        const pageItems = state.pagedItems[state.currentPage] || [];
-                        if (pageItems[idx]) {
-                            pageItems[idx]._deleted = this.checked;
-                        }
-                    });
-                });
-            }
-
-            function renderMessageDetail() {
-                const container = document.getElementById('messageDetail');
-                if (!container) return;
-                container.style.display = state.selected ? '' : 'none';
-                if (!state.selected) return;
-
-                const sel = state.selected;
-                const headerEl = document.getElementById('detailHeader');
-                if (headerEl) {
-                    headerEl.innerHTML = `<h5 class="pt-2">
-                        <a href="javascript:;" id="btnCloseDetail"><?php echo xlt('Conversation from'); ?></a>
-                        <strong>${escapeHtml(sel.sender_name)}</strong> <?php echo xlt('regarding'); ?> <strong>${escapeHtml(sel.title)}</strong> <?php echo xlt('on'); ?> &lt;${escapeHtml(sel.date)}&gt;
-                    </h5>`;
-                    document.getElementById('btnCloseDetail').addEventListener('click', closeMessage);
-                }
-
-                // Render chained messages
-                const chainTbody = document.getElementById('chainTbody');
-                if (!chainTbody) return;
-                const chainId = sel.mail_chain;
-                const chained = isNaN(chainId) ? state.allItems : state.allItems.filter(item => item.mail_chain == chainId);
-
-                let html = '';
-                chained.forEach(item => {
-                    const isSelected = sel.id === item.id;
-                    const showReply = sel.sender_id !== state.cUserId && isSelected;
-                    const showForward = isSelected && sel.sender_id !== state.cUserId && !state.isPortal;
-                    const showDelete = !state.isTrash && isSelected;
-
-                    html += `<tr>
-                        <td role="button">
-                            <span class="col-sm px-1"><span>${escapeHtml(item.date)}</span></span>
-                            <span class="col-sm"><span>${escapeHtml(item.message_status)}</span></span>
-                            <span class="col-sm px-1"><span>${escapeHtml(item.sender_name)} to ${escapeHtml(item.recipient_name)}</span></span>
-                            <span class="col-sm-1"><span>${escapeHtml(item.title)}</span></span>`;
-
-                    if (!isSelected) {
-                        html += `<span class="col-sm px-1"><span>${limitTo(htmlToText(item.body), 35)}</span></span>`;
-                    }
-
-                    html += `<span class='btn-group float-right m-0'>`;
-                    if (showReply) {
-                        html += `<button class="btn btn-primary btn-small btn-compose-reply" data-toggle="modal" data-mode="reply" data-noteid="${escapeHtml(sel.id)}" data-whoto="${escapeHtml(sel.sender_id)}" data-mtitle="${escapeHtml(sel.title)}" data-username="${escapeHtml(sel.sender_name)}" data-mailchain="${escapeHtml(sel.mail_chain)}" data-target="#modalCompose"><i class="fa fa-reply"></i></button>`;
-                    }
-                    if (showForward) {
-                        html += `<button class="btn btn-primary btn-small btn-compose-fwd" data-toggle="modal" data-mode="forward" data-noteid="${escapeHtml(sel.id)}" data-whoto="${escapeHtml(sel.sender_id)}" data-mtitle="${escapeHtml(sel.title)}" data-username="${escapeHtml(sel.sender_name)}" data-mailchain="${escapeHtml(sel.mail_chain)}" data-target="#modalCompose"><i class="fa fa-share"></i></button>`;
-                    }
-                    if (showDelete) {
-                        html += `<button class="btn btn-small btn-primary btn-delete-item" data-item-id="${escapeHtml(sel.id)}" title="<?php echo xla('Archive this message'); ?>"><i class="fa fa-trash fa-1x"></i></button>`;
-                    }
-                    html += `</span>`;
-
-                    if (isSelected) {
-                        html += `<div class='col jumbotron jumbotron-fluid my-3 p-1 bg-light text-dark rounded border border-info'><span>${renderMessageBody(sel.body)}</span></div>`;
-                    }
-                    html += `</td></tr>`;
-                });
-                chainTbody.innerHTML = html;
-
-                // Bind reply/forward buttons to open modal
-                chainTbody.querySelectorAll('.btn-compose-reply, .btn-compose-fwd').forEach(btn => {
-                    btn.addEventListener('click', function () {
-                        $('#modalCompose').data('relatedTarget', this);
-                        $('#modalCompose').modal('show');
-                    });
-                });
-                // Bind delete
-                chainTbody.querySelectorAll('.btn-delete-item').forEach(btn => {
-                    btn.addEventListener('click', () => deleteItem(state.selected));
-                });
-            }
-
-            function renderPaging() {
-                const pagingEl = document.getElementById('pagingControls');
-                if (!pagingEl) return;
-                pagingEl.style.display = state.selected ? 'none' : '';
-
-                const currentPageItems = state.pagedItems[state.currentPage] || [];
-                const startNum = (state.itemsPerPage * state.currentPage) + 1;
-                const endNum = (state.itemsPerPage * state.currentPage) + currentPageItems.length;
-
-                const infoEl = document.getElementById('pagingInfo');
-                if (infoEl) {
-                    infoEl.innerHTML = `<strong>${startNum}</strong>~<strong>${endNum}</strong> of <strong>${state.items.length}</strong>`;
-                }
-
-                const btnGroup = document.getElementById('pagingButtons');
-                if (btnGroup) {
-                    btnGroup.style.display = state.items.length > state.itemsPerPage ? '' : 'none';
-                }
-            }
-
-            // --- Compose modal setup ---
-
-            function setupComposeModal(e) {
-                state.compose = { task: 'add' };
-                $('#inputBody').summernote('destroy');
-
-                // Get mode from data-mode attribute
-                const relatedTarget = $(e.currentTarget).data('relatedTarget') || e.relatedTarget;
-                const mode = relatedTarget ? $(relatedTarget).attr('data-mode') : 'add';
-                state.compose.task = mode;
-
-                // Populate the "Refer to Message" panel with the sanitized
-                // body of the currently selected message when replying or
-                // forwarding. This replaces the former Angular ng-bind-html.
-                const referMsgEl = document.getElementById('referMsg');
-                const referLabelEl = document.getElementById('referLabel');
-                const referMsgIdEl = document.getElementById('referMsgId');
-                const showRefer = (mode === 'forward' || mode === 'reply') && state.selected && state.selected.mail_chain;
-                if (referMsgEl) {
-                    if (showRefer) {
-                        referMsgEl.innerHTML = renderMessageBody(state.selected.body);
-                        referMsgEl.style.display = '';
-                    } else {
-                        referMsgEl.innerHTML = '';
-                        referMsgEl.style.display = 'none';
-                    }
-                }
-                if (referLabelEl) {
-                    referLabelEl.style.display = showRefer ? '' : 'none';
-                }
-                if (referMsgIdEl && showRefer) {
-                    referMsgIdEl.textContent = state.selected.id;
-                }
-
-                if (mode === 'forward') {
-                    $('#modalCompose .modal-header .modal-title').html("Forward Message");
-                    const recipId = $(relatedTarget).attr('data-whoto');
-                    const title = $(relatedTarget).attr('data-mtitle');
-                    const uname = $(relatedTarget).attr('data-username');
-                    $('#selForwardto').prop('disabled', false).show();
-                    $('#selSendto').hide();
-                    $('#title').prop('disabled', false);
-                    state.compose.title = title;
-                    state.compose.selrecip = recipId;
-                    state.compose.recipient_name = uname;
-                    state.compose.recipient_id = recipId;
-
-                    state.authrecips.forEach(o => {
-                        if (o.userid === recipId) {
-                            state.compose.pid = o.pid;
-                        }
-                    });
-                    const referText = referMsgEl ? (referMsgEl.textContent || referMsgEl.innerText || '') : '';
-                    const fmsg = '\n\n\n> ' + state.xLate.fwd + title + ' by ' + uname + '\n> ' + referText;
-                    $('#finputBody').val(fmsg).show();
-                    $('#inputBody').hide();
-                    state.compose.noteid = $(relatedTarget).attr('data-noteid');
-                    updateComposeForm();
-                } else if (mode === 'reply') {
-                    $('#inputBody').show().summernote({
-                        focus: true,
-                        height: '225px',
-                        width: '100%',
-                        tabsize: 4,
-                        disableDragAndDrop: true,
-                        dialogsInBody: true,
-                        dialogsFade: true
-                    });
-                    $('#finputBody').hide();
-                    $('#selForwardto').hide();
-                    $('#selSendto').show();
-                    $('#modalCompose .modal-header .modal-title').html(<?php echo xlj("Compose Reply Message"); ?>);
-                    let chain = $(relatedTarget).attr('data-mailchain');
-                    if (chain === '0') {
-                        chain = $(relatedTarget).attr('data-noteid');
-                    }
-                    const recipId = $(relatedTarget).attr('data-whoto');
-                    const title = $(relatedTarget).attr('data-mtitle');
-                    const uname = $(relatedTarget).attr('data-username');
-                    $('#selSendto').val(recipId);
-                    $('#title').val(title);
-                    state.compose.title = title;
-                    state.compose.selrecip = recipId;
-                    state.compose.recipient_name = uname;
-                    state.compose.recipient_id = recipId;
-                    state.compose.noteid = chain;
-                    updateComposeForm();
-                } else {
-                    $('#inputBody').show().summernote({
-                        width: '100%',
-                        focus: true,
-                        height: '375px',
-                        tabsize: 4,
-                        disableDragAndDrop: true,
-                        dialogsInBody: true,
-                        dialogsFade: true,
-                        popover: {
-                            image: [],
-                            link: [],
-                            air: []
-                        }
-                    });
-                    $('#finputBody').hide();
-                    $('#selForwardto').hide();
-                    $('#selSendto').show().prop('disabled', false);
-                    $('#modalCompose .modal-header .modal-title').html(<?php echo xlj("Compose New Message"); ?>);
-                    state.compose.task = 'add';
-                    $('#selSendto').prop('disabled', false);
-                    $('#title').prop('disabled', false);
-                    updateComposeForm();
-                }
-            }
-
-            function updateComposeForm() {
-                // Sync hidden fields with compose state
-                $('#hiddenTask').val(state.compose.task || 'add');
-                $('#hiddenNoteid').val(state.compose.noteid || '');
-                if (state.compose.title) {
-                    $('#title').val(state.compose.title);
-                }
-                if (state.compose.selrecip) {
-                    $('#selSendto').val(state.compose.selrecip);
-                    $('#selForwardto').val(state.compose.selrecip);
-                }
-            }
-
-            function submitForm(e) {
-                // Re-enable title for submit
-                $('#title').prop('disabled', false);
-                $('#selSendto').prop('disabled', false);
-
-                const task = state.compose.task || 'add';
-                const selrecip = $('#selSendto').val() || state.compose.selrecip;
-
-                // Confirm when sending to yourself
-                if (selrecip === state.cUserId) {
-                    if (!confirm(state.xLate.confirm.err)) {
-                        e.preventDefault();
-                        return false;
-                    }
-                }
-
-                // Populate hidden inputs so the natural form submission carries
-                // the CSRF token and all metadata expected by handle_note.php.
-                document.getElementById('csrf_token_form').value = state.csrf;
-                document.getElementById('hiddenTask').value = task;
-                document.getElementById('hiddenSenderName').value = state.userproper;
-
-                if (task === 'forward') {
-                    const $fwd = $('#selForwardto option:selected');
-                    document.getElementById('hiddenSenderId').value = $fwd.val() || '';
-                    document.getElementById('hiddenSenderName').value = $fwd.text() || state.userproper;
-                    document.getElementById('hiddenRecipientId').value = state.compose.recipient_id || '';
-                    document.getElementById('hiddenRecipientName').value = state.compose.recipient_name || '';
-                    document.getElementById('hiddenPid').value = state.compose.pid || '';
-                    document.getElementById('hiddenNoteid').value = state.compose.noteid || '';
-                    // finputBody textarea already has its value via its name="inputBody"
-                } else {
-                    document.getElementById('hiddenSenderId').value = state.cUserId;
-                    const recipName = $('#selSendto option:selected').text() || state.compose.recipient_name || '';
-                    document.getElementById('hiddenRecipientName').value = recipName;
-                    document.getElementById('hiddenRecipientId').value = selrecip || '';
-                    document.getElementById('hiddenNoteid').value = state.compose.noteid || '';
-                    // Copy Summernote contents into the hidden finputBody (which has name="inputBody")
-                    document.getElementById('finputBody').value = $('#inputBody').summernote('code');
-                }
-
-                // Set replyid for message threading (replies and forwards).
-                // Matches the original AngularJS ng-value='selected.reply_mail_chain'.
-                document.getElementById('replyid').value = state.selected?.reply_mail_chain || '';
-
-                // Let the form submit naturally with all values now populated
-                return true;
-            }
-
-            // --- Data init ---
-
-            async function initData() {
-                try {
-                    await Promise.all([
-                        getSentMessages(),
-                        getAllMessages(),
-                        getDeletedMessages()
-                    ]);
-                    state.isInit = true;
-                    state.errorLoadingMessages = false;
-                } catch (e) {
-                    state.errorLoadingMessages = true;
-                    state.isInit = true;
-                }
-                selectInbox();
-            }
-
-            // --- DOM Ready ---
-
-            document.addEventListener('DOMContentLoaded', function () {
-                // Nav bindings
-                document.getElementById('navInbox').addEventListener('click', e => { e.preventDefault(); selectInbox(); });
-                document.getElementById('navSent').addEventListener('click', e => { e.preventDefault(); selectSent(); });
-                document.getElementById('navAll').addEventListener('click', e => { e.preventDefault(); selectAll(); });
-                document.getElementById('navTrash').addEventListener('click', e => { e.preventDefault(); selectTrash(); });
-
-                // Toolbar bindings
-                document.getElementById('btnReadAll').addEventListener('click', e => { e.preventDefault(); readAll(); });
-                document.getElementById('btnBatchDelete').addEventListener('click', e => { e.preventDefault(); batchDelete(); });
-
-                // Paging
-                document.getElementById('btnPrev').addEventListener('click', e => { e.preventDefault(); prevPage(); });
-                document.getElementById('btnNext').addEventListener('click', e => { e.preventDefault(); nextPage(); });
-
-                // Compose modal
-                $('#modalCompose').on('hidden.bs.modal', function () {
-                    window.location.reload();
-                });
-                $('#modalCompose').on('show.bs.modal', setupComposeModal);
-
-                // Form submit
-                const form = document.getElementById('fcompose');
-                if (form) {
-                    form.addEventListener('submit', submitForm);
-                }
-
-                // Populate recipient selects
-                const selSendto = document.getElementById('selSendto');
-                const selForwardto = document.getElementById('selForwardto');
-                state.authrecips.forEach(recip => {
-                    const opt = document.createElement('option');
-                    opt.value = recip.userid;
-                    opt.textContent = recip.username;
-                    selSendto.appendChild(opt);
-
-                    if (recip.type === 'user') {
-                        const opt2 = document.createElement('option');
-                        opt2.value = recip.userid;
-                        opt2.textContent = recip.username;
-                        selForwardto.appendChild(opt2);
-                    }
-                });
-
-                // Portal-only visibility
-                if (state.isPortal) {
-                    document.querySelectorAll('[data-show-dashboard]').forEach(el => el.style.display = 'none');
-                }
-
-                // Initialize data
-                initData();
-            });
-
-            <?php
-            if ($showSMS) {
-                $globalsBag->getKernel()->getEventDispatcher()->dispatch(new SendSmsEvent($pid), SendSmsEvent::JAVASCRIPT_READY_SMS_POST);
-            }
-            ?>
-        })();
+        };
     </script>
+    <script src="<?php echo attr($webRoot); ?>/portal/messaging/js/messages.js?v=<?php echo attr($jsVersion); ?>"></script>
+    <?php if ($showSMS) { ?>
+    <script>
+        <?php $globalsBag->getKernel()->getEventDispatcher()->dispatch(new SendSmsEvent($pid), SendSmsEvent::JAVASCRIPT_READY_SMS_POST); ?>
+    </script>
+    <?php } ?>
     <div class="container-fluid" id='main'>
         <div class='my-3'>
             <h2><i class='fa fa-envelope w-auto h-auto mr-2'></i><?php echo xlt('Secure Messaging'); ?></h2>
